- allow higher resolution layers for Maxent and ALOC - speedup index creation
git-svn-id: http://alageospatialportal.googlecode.com/svn/trunk/alaspatial@345 73def43e-0e3b-11df-a5af-b38b39d06e06

+++ src/main/java/org/ala/spatial/util/Zipper.java
@@ -99,7 +99,7 @@ public class Zipper {
         System.out.println("v: " + g.maxval);
         System.out.println("size: " + g.xres + " x " + g.yres); 
 
-        double[] dv = g.getGrid();
+        float[] dv = g.getGrid();
 
         System.out.println("values:\n" + dv.length);
         for (int i=0; i<dv.length; i++) {
